backend/requirements.txt
flask
openai
python-dotenv
streamlit
Pillow
requests
transformers
torch
google-generativeai